feat: 걸음수 등록 API

+++ mobile/app/src/main/java/com/secui/healthone/data/WalkData.kt
@@ -3,5 +3,5 @@ package com.secui.healthone.data
 data class WalkData(
     val userNo: Int,
     val stepCount: Int,
-    val distance: Float
+    val distance: Double
 )
